Implement strong exception guarantee
reallocate now has a strong exception guarantee. On exception it will
revert the state of the vector to before the function was called. This
behavior is unspecified for objects that are noncopyable and can throw on
move.
A default Allocator template argument was added (which does nothing at the
moment.)
The noexcept specifier was added to all acceptable functions.
An object copying constructor, swap method, and data method were added.